refactor: configure playwright static server
