+++ playwright.config.js
@@ -1,0 +1,19 @@
+const { defineConfig, devices } = require('@playwright/test');
+
+module.exports = defineConfig({
+  testDir: 'tests',
+  use: {
+    baseURL: 'http://127.0.0.1:4173',
+  },
+  projects: [
+    {
+      name: 'chromium',
+      use: { ...devices['Desktop Chrome'] },
+    },
+  ],
+  webServer: {
+    command: 'node tests/static-server.js',
+    port: 4173,
+    reuseExistingServer: !process.env.CI,
+  },
+});
